We can now process all observations passed to a given simulation.
Scheduler.run() was not exiting after the rest of the simulation had
finished; there is now a 'hard' exit by checking if the telescope has
no more observations.

WorkflowPlans now have a WorkflowPlanStatus
FIFOScheduler now checks EST + workflow.start_time to determine if a
task is ready to be scheduled.