Update where planning occurs in simulation.
The reasoning behind this is to provide buffer data to the planner at generation time;
the planner then stores the estimated time the workflow itself is meant to start,
due to a) when the observation finishes, plus b) the estimated time delay between the
transition of data from hot to cold buffer.

From this, the scheduler may inspect the workflow and raise a 'delay' flag in the even
that workflows are delayed on submission to the scheduler.