Scheduler now accurately allocates;
allocation decisions are now made in bulk, as opposed to one allocation
*max* each timestep.

Steps have been taken to improve the use of the cluster to ensure
that mutable objects are not accessed outside of the cluster.

Planner behaviour has been modified to provide all resources to all
observations that are being planned for.