Scheduler and scheduling policies based on plans have been updated to ensure data-transfer time happens within the simulation, rather than as a product of the estimated start time (EST) as determined in the plan.
This allows a simplified dynamic-schedule-from-plan approach; improves the
accuracy of the model; and paves the way for a batch-scheduling approach that
does not require additional work within the Scheduler.

Tasks are now in control of their compute duration _and_ their their data-transfer delays.
These currently do not have a delay feature, but are computed based on data cost established
from the workflow structure.

Future work on the batch scheduler will require only the use of a sorting 'queue',
and the allocation of estimated start times that are placeholder values for
the purpose of sorting in the scheduler.