Reverted to raising a RunTime error on the cluster if a double-allocation is made.
This means that if developers want to use the Cluster object methods themselves, the cluster will error out as when double allocation is made. Note: it is not possible to catch the RuntimeError at the time allocate_tasks is called, because the function is a generator and won't properly be invoked until it is run on the environment.