BugFix: Buffer hangs when residual data < datatransfer
This error manifests due to the check in receive_transfer or transfer_observation.
We want to check if data == 0, because if data is <0, then we've taken up more data
than we should have, and if it's >0 there's still more to go.
However, we weren't check for residual data sizes less than the transfer rate, so it was never reaching that check and thus observation was always stored in ['transfer'].