#53: Add use_task_data and use_edge_data parameters for simulation.
This is to simplify observation plan generation and instead pass the onus of experiment combinations onto the simulator.

These changes also include some polishing refactoring of the Task and Scheduler classes and their respective tests. Future work to refactor and simplify/move redundancies has been identified and added as #TODOs; I'll add this to a GitHub issue once this has been self-reviewed.